Recursion

+++ heaps/closest_stars.cc
@@ -6,7 +6,7 @@ namespace algo {
 
 std::vector<Star> FindClosestStars(std::vector<Star>::const_iterator begin,
                                    const std::vector<Star>::const_iterator& end,
-                                   int32_t k) {
+                                   size_t k) {
   // We want max_heap
   std::priority_queue<Star, std::vector<Star>, std::less<>> max_heap;
